requirements/base.txt
Django==1.8.3
psycopg2==2.6.1
django-pgjsonb==0.0.16
Pillow==2.9.0
djangorestframework==3.2.0
